+++ project/src/Styles/Minesweeper.scss
@@ -10,8 +10,8 @@
         display: flex;
         justify-content: center;
         align-items: center;
-        width: 80px;
-        height: 80px;
+        width: 45px;
+        height: 45px;
         text-align: center;
         border: 2px solid black;
         font-weight: 800;
